pyproject.toml
[project]
name = "ats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "langchain-openai>=1.1.7",
    "pdf2image>=1.17.0",
    "pydantic>=2.12.5",
    "pypdf>=6.6.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
]
